@@ -6,7 +6,7 @@ version: 0.1.1
 homepage: https://github.com/huitingyeong/bluetooth_enable
 
 environment:
-  sdk: ">=2.1.0 <3.0.0"
+  sdk: '>=2.12.0 <3.0.0'
 
 dependencies:
   flutter:
